requirements.txt

streamlit>=1.28.0
streamlit-chat>=0.1.1
google-generativeai>=0.3.0
pandas>=1.5.0